Remove PossibleCombination and tweak Shuffler
